\name{rank_val}
\alias{rank_val}
\title{Rank value search}
\description{
  \code{rank_val} look for the value of a given rank in a vector.
}
\usage{
  rank_val(target,rk)

}
\arguments{
  \item{target}{numeric vector, data for which the rank value will be look for}

  \item{rk}{numeric, the rank to look for}

}

\value{
  \code{rank_val}, returns a the value with rank rk in the vector.
}
\details{
  \code{rank_val}  look for the value of a given rank in a vector.
}

